refactor: notification panel components and utility functions
