+++ ai-fusion-video-web/components/dashboard/notification-panel/detail.tsx
@@ -1,0 +1,644 @@
+"use client";
+
+import { useCallback, useEffect, useRef, useState, type ReactNode } from "react";
+import { createPortal } from "react-dom";
+import { AnimatePresence, motion } from "framer-motion";
+import {
+  Ban,
+  CheckCircle2,
+  ChevronRight,
+  Clock,
+  Loader2,
+  MessageSquare,
+  X,
+  XCircle,
+} from "lucide-react";
+import {
+  listConversations,
+  listMessages,
+  type AgentConversation,
+  type AgentMessage,
+} from "@/lib/api/ai-assistant";
+import { PIPELINE_AGENT_TYPES } from "@/lib/api/ai-pipeline";
+import {
+  usePipelineStore,
+  type PipelineTask,
+} from "@/lib/store/pipeline-store";
+import { cn } from "@/lib/utils";
+import {
+  getAgentTypeName,
+  getToolDisplayName,
+} from "./constants";
+import {
+  messagesToTimeline,
+  resolveHistoryErrorMessage,
+} from "./history";
+import { ElapsedText, useSmartScroll } from "./hooks";
+import { MessageTimeline } from "./timeline";
+import {
+  formatDatetime,
+  formatElapsed,
+  formatTimestamp,
+  getElapsedStr,
+} from "./utils";
+
+function PipelineDetailPanel({ task }: { task: PipelineTask }) {
+  const [idleTimelineLength, setIdleTimelineLength] = useState<number | null>(null);
+  const timelineLength = task.state.timeline.length;
+  const isIdle = task.status === "running" && idleTimelineLength === timelineLength;
+  const canCancel = task.status === "running" && task.cancellable !== false;
+  const timelineRef = useSmartScroll([task.state.timeline, isIdle], task.status === "running");
+
+  useEffect(() => {
+    if (task.status !== "running") return;
+    const timer = setTimeout(() => {
+      setIdleTimelineLength(timelineLength);
+    }, 2000);
+    return () => clearTimeout(timer);
+  }, [timelineLength, task.status]);
+
+  const statusText = {
+    running: "运行中",
+    done: "已完成",
+    error: "出错",
+    cancelled: "已取消",
+  };
+
+  const statusColor = {
+    running: "text-blue-400",
+    done: "text-green-400",
+    error: "text-destructive",
+    cancelled: "text-muted-foreground",
+  };
+
+  return (
+    <div className="flex flex-col h-full">
+      <div className="px-4 py-3 border-b border-border/20 shrink-0 flex items-start justify-between gap-2">
+        <div className="min-w-0">
+          <h4 className="text-sm font-semibold truncate">{task.label}</h4>
+          <p className={cn("text-xs mt-0.5", statusColor[task.status])}>
+            {statusText[task.status]} · <ElapsedText task={task} />
+            <span className="text-muted-foreground/50 ml-1">
+              启动于 {formatTimestamp(task.createdAt)}
+            </span>
+          </p>
+        </div>
+        {canCancel && (
+          <button
+            onClick={() => usePipelineStore.getState().cancelPipeline(task.id)}
+            className="shrink-0 flex items-center gap-1.5 px-2.5 py-1.5 rounded-lg text-[11px] font-medium border border-destructive/20 text-destructive/70 hover:bg-destructive/10 hover:text-destructive hover:border-destructive/40 transition-colors"
+            title="停止工作流"
+          >
+            <Ban className="h-3 w-3" />
+            停止
+          </button>
+        )}
+      </div>
+
+      <div ref={timelineRef} className="flex-1 min-h-0 overflow-y-auto p-4 space-y-3">
+        <MessageTimeline
+          reasoningText={task.state.reasoningText}
+          reasoningDurationMs={task.state.reasoningDurationMs}
+          timeline={task.state.timeline}
+          streaming={task.status === "running"}
+          error={task.state.error}
+        />
+
+        <AnimatePresence>
+          {isIdle && task.status === "running" && (
+            <motion.div
+              initial={{ opacity: 0, y: 4 }}
+              animate={{ opacity: 1, y: 0 }}
+              exit={{ opacity: 0, y: -4 }}
+              transition={{ duration: 0.25 }}
+              className="flex items-center gap-2 px-3 py-2 rounded-lg bg-primary/5 border border-primary/10"
+            >
+              <Loader2 className="h-3 w-3 animate-spin text-primary/60" />
+              <span className="text-xs text-primary/70 font-medium">AI 全力处理中…</span>
+            </motion.div>
+          )}
+        </AnimatePresence>
+      </div>
+    </div>
+  );
+}
+
+function HistoryDetailPanel({
+  conversation,
+}: {
+  conversation: AgentConversation;
+}) {
+  const [messageState, setMessageState] = useState<{
+    conversationId: string;
+    messages: AgentMessage[];
+  } | null>(null);
+  const loading = messageState?.conversationId !== conversation.conversationId;
+  const messages =
+    messageState?.conversationId === conversation.conversationId
+      ? messageState.messages
+      : [];
+
+  useEffect(() => {
+    let cancelled = false;
+    listMessages(conversation.conversationId)
+      .then((nextMessages) => {
+        if (!cancelled) {
+          setMessageState({
+            conversationId: conversation.conversationId,
+            messages: nextMessages,
+          });
+        }
+      })
+      .catch((err) => {
+        console.error(err);
+        if (!cancelled) {
+          setMessageState({
+            conversationId: conversation.conversationId,
+            messages: [],
+          });
+        }
+      });
+
+    return () => {
+      cancelled = true;
+    };
+  }, [conversation.conversationId]);
+
+  const isPipeline =
+    conversation.category === "task" ||
+    conversation.category === "pipeline" ||
+    (conversation.agentType != null &&
+      (PIPELINE_AGENT_TYPES as readonly string[]).includes(conversation.agentType));
+  const displayMessages = isPipeline
+    ? messages.filter((message) => message.role !== "user")
+    : messages;
+  const firstAssistant = displayMessages.find(
+    (message) =>
+      message.role === "assistant" &&
+      !message.parentToolCallId &&
+      message.reasoningContent
+  );
+  const timeline = messagesToTimeline(displayMessages);
+  const historyError = resolveHistoryErrorMessage(conversation, displayMessages);
+
+  return (
+    <div className="flex flex-col h-full">
+      <div className="px-4 py-3 border-b border-border/20 shrink-0">
+        <h4 className="text-sm font-semibold truncate">{conversation.title}</h4>
+        <div className="flex items-center gap-2 mt-0.5 flex-wrap">
+          {conversation.agentType && (
+            <span className="text-[10px] px-1.5 py-0.5 rounded bg-blue-500/10 text-blue-400 font-medium">
+              {getAgentTypeName(conversation.agentType)}
+            </span>
+          )}
+          {conversation.createTime && conversation.lastMessageTime && (() => {
+            const duration =
+              new Date(conversation.lastMessageTime).getTime() -
+              new Date(conversation.createTime).getTime();
+            return duration > 0 ? (
+              <span className="text-xs text-muted-foreground">
+                耗时 {formatElapsed(duration)}
+              </span>
+            ) : null;
+          })()}
+          <span className="text-xs text-muted-foreground">
+            {formatDatetime(conversation.createTime)}
+          </span>
+        </div>
+      </div>
+
+      <div className="flex-1 min-h-0 overflow-y-auto p-4 space-y-3">
+        {loading ? (
+          <div className="flex items-center justify-center py-8">
+            <Loader2 className="h-5 w-5 animate-spin text-muted-foreground" />
+          </div>
+        ) : displayMessages.length === 0 ? (
+          <p className="text-sm text-muted-foreground text-center py-8">暂无消息记录</p>
+        ) : (
+          <MessageTimeline
+            reasoningText={firstAssistant?.reasoningContent || undefined}
+            reasoningDurationMs={firstAssistant?.reasoningDurationMs || undefined}
+            timeline={timeline}
+            streaming={false}
+            error={historyError}
+          />
+        )}
+      </div>
+    </div>
+  );
+}
+
+function TaskListItem({
+  label,
+  subtitle,
+  icon,
+  selected,
+  onClick,
+}: {
+  label: string;
+  subtitle: ReactNode;
+  icon: ReactNode;
+  selected: boolean;
+  onClick: () => void;
+}) {
+  return (
+    <button
+      onClick={onClick}
+      className={cn(
+        "w-full flex items-center gap-2 px-2.5 py-2 rounded-lg text-left transition-colors mb-0.5",
+        selected ? "bg-foreground/10" : "hover:bg-foreground/5"
+      )}
+    >
+      {icon}
+      <div className="flex-1 min-w-0">
+        <p className="text-xs font-medium truncate">{label}</p>
+        <p className="text-[10px] text-muted-foreground truncate">{subtitle}</p>
+      </div>
+    </button>
+  );
+}
+
+export function PipelineTaskCard({ task }: { task: PipelineTask }) {
+  const { setPanelExpanded, setExpandedTaskId } = usePipelineStore();
+  const isRunning = task.status === "running";
+
+  const statusIcon = {
+    running: <Loader2 className="h-3.5 w-3.5 animate-spin text-blue-400 shrink-0" />,
+    done: <CheckCircle2 className="h-3.5 w-3.5 text-green-400 shrink-0" />,
+    error: <XCircle className="h-3.5 w-3.5 text-destructive shrink-0" />,
+    cancelled: <Ban className="h-3.5 w-3.5 text-muted-foreground shrink-0" />,
+  };
+
+  const statusText = {
+    running: "运行中",
+    done: "已完成",
+    error: "出错",
+    cancelled: "已取消",
+  };
+
+  const getLatestActivity = () => {
+    const timeline = task.state.timeline;
+    const isTaskStream = task.cancellable === false;
+    for (let index = timeline.length - 1; index >= 0; index--) {
+      const item = timeline[index];
+      if (item.type === "reasoning") {
+        return "AI 正在思考…";
+      }
+      if (item.type === "tool") {
+        return item.status === "calling"
+          ? `正在${getToolDisplayName(item.name)}…`
+          : `${getToolDisplayName(item.name)} 已完成`;
+      }
+      if (item.type === "content") {
+        if (isTaskStream) {
+          return task.status === "running" ? "任务执行中…" : "已记录任务结果";
+        }
+        return task.status === "running" ? "AI 正在输出…" : "已生成回复";
+      }
+    }
+
+    if (task.state.reasoningText) return "AI 正在思考…";
+    return isTaskStream ? "任务准备中…" : "准备中…";
+  };
+
+  const handleOpenDetail = () => {
+    setExpandedTaskId(task.id);
+    setPanelExpanded(true);
+  };
+
+  return (
+    <motion.div
+      layout
+      initial={{ opacity: 0, y: -8 }}
+      animate={{ opacity: 1, y: 0 }}
+      exit={{ opacity: 0, y: -8, height: 0 }}
+      className={cn(
+        "rounded-xl border overflow-hidden transition-colors",
+        isRunning
+          ? "border-blue-500/20 bg-blue-500/5"
+          : task.status === "done"
+            ? "border-green-500/20 bg-green-500/5"
+            : task.status === "error"
+              ? "border-destructive/20 bg-destructive/5"
+              : "border-border/20 bg-muted/20"
+      )}
+    >
+      <div
+        className="flex items-center gap-2 px-3 py-2.5 cursor-pointer hover:bg-black/5 dark:hover:bg-white/5 transition-colors"
+        onClick={handleOpenDetail}
+      >
+        {statusIcon[task.status]}
+        <div className="flex-1 min-w-0">
+          <p className="text-xs font-medium truncate">{task.label}</p>
+          <p className="text-[10px] text-muted-foreground truncate">
+            {isRunning ? getLatestActivity() : statusText[task.status]}
+            {" · "}
+            <ElapsedText task={task} />
+          </p>
+        </div>
+        <div className="flex items-center shrink-0">
+          <ChevronRight className="h-3 w-3 text-muted-foreground" />
+        </div>
+      </div>
+    </motion.div>
+  );
+}
+
+type SelectedItem =
+  | { type: "pipeline"; taskId: string }
+  | { type: "history"; conversation: AgentConversation };
+
+export function ExpandedPanel({ onClose }: { onClose: () => void }) {
+  const { tasks, clearCompleted, expandedTaskId } = usePipelineStore();
+  const listEndRef = useRef<HTMLDivElement>(null);
+
+  const [selected, setSelected] = useState<SelectedItem | null>(() => {
+    if (expandedTaskId) {
+      const target = tasks.find((task) => task.id === expandedTaskId);
+      if (target) return { type: "pipeline", taskId: target.id };
+    }
+    const running = tasks.find((task) => task.status === "running");
+    return running ? { type: "pipeline", taskId: running.id } : null;
+  });
+  const [conversations, setConversations] = useState<AgentConversation[]>([]);
+  const [historyPage, setHistoryPage] = useState(1);
+  const [historyTotal, setHistoryTotal] = useState(0);
+  const [historyLoading, setHistoryLoading] = useState(false);
+  const pageSize = 20;
+
+  const hasMore = conversations.length < historyTotal;
+
+  const loadHistory = useCallback(
+    async (page: number, append = false) => {
+      setHistoryLoading(true);
+      try {
+        const result = await listConversations({ pageNo: page, pageSize });
+        const currentConversationIds = new Set(
+          usePipelineStore
+            .getState()
+            .tasks.filter((task) => task.state.conversationId)
+            .map((task) => task.state.conversationId!)
+        );
+        const filtered = result.list.filter(
+          (conversation) =>
+            conversation.status !== "running" &&
+            !currentConversationIds.has(conversation.conversationId)
+        );
+
+        if (append) {
+          setConversations((prev) => [...prev, ...filtered]);
+        } else {
+          setConversations(filtered);
+        }
+
+        setHistoryTotal(result.total - (result.list.length - filtered.length));
+        setHistoryPage(page);
+      } catch (err) {
+        console.error("加载历史对话失败:", err);
+      } finally {
+        setHistoryLoading(false);
+      }
+    },
+    [pageSize]
+  );
+
+  useEffect(() => {
+    loadHistory(1);
+  }, [loadHistory]);
+
+  const prevTasksRef = useRef(tasks);
+  useEffect(() => {
+    const prevTasks = prevTasksRef.current;
+    prevTasksRef.current = tasks;
+
+    const justFinished = tasks.some((task) => {
+      if (task.status === "running") return false;
+      const prev = prevTasks.find((previousTask) => previousTask.id === task.id);
+      return prev && prev.status === "running";
+    });
+
+    if (justFinished) {
+      const timer = setTimeout(() => {
+        loadHistory(1);
+        clearCompleted();
+      }, 1500);
+      return () => clearTimeout(timer);
+    }
+  }, [tasks, loadHistory, clearCompleted]);
+
+  useEffect(() => {
+    if (!listEndRef.current || !hasMore) return;
+
+    const observer = new IntersectionObserver(
+      (entries) => {
+        if (entries[0].isIntersecting && !historyLoading && hasMore) {
+          loadHistory(historyPage + 1, true);
+        }
+      },
+      { threshold: 0.1 }
+    );
+    observer.observe(listEndRef.current);
+    return () => observer.disconnect();
+  }, [hasMore, historyLoading, historyPage, loadHistory]);
+
+  useEffect(() => {
+    const handleKeyDown = (event: KeyboardEvent) => {
+      if (event.key === "Escape") onClose();
+    };
+    document.addEventListener("keydown", handleKeyDown);
+    return () => document.removeEventListener("keydown", handleKeyDown);
+  }, [onClose]);
+
+  const selectedPipelineTask =
+    selected?.type === "pipeline"
+      ? tasks.find((task) => task.id === selected.taskId)
+      : null;
+  const selectedConversation =
+    selected?.type === "history" ? selected.conversation : null;
+  const runningTasks = tasks.filter((task) => task.status === "running");
+  const completedTasks = tasks.filter((task) => task.status !== "running");
+
+  return createPortal(
+    <>
+      <motion.div
+        initial={{ opacity: 0 }}
+        animate={{ opacity: 1 }}
+        exit={{ opacity: 0 }}
+        className="fixed inset-0 z-60 bg-black/40 backdrop-blur-sm"
+        onClick={onClose}
+      />
+      <motion.div
+        initial={{ opacity: 0, scale: 0.95, y: 20 }}
+        animate={{ opacity: 1, scale: 1, y: 0 }}
+        exit={{ opacity: 0, scale: 0.95, y: 20 }}
+        transition={{ duration: 0.2, ease: "easeOut" }}
+        className="fixed z-61 left-1/2 top-1/2 -translate-x-1/2 -translate-y-1/2 w-[900px] max-w-[92vw] h-[75vh] rounded-2xl border border-border/40 bg-card/98 backdrop-blur-xl shadow-2xl shadow-black/30 flex flex-col overflow-hidden"
+      >
+        <div className="flex items-center justify-between px-5 py-3.5 border-b border-border/20 shrink-0">
+          <h3 className="text-base font-semibold">AI 任务中心</h3>
+          <div className="flex items-center gap-2">
+            {completedTasks.length > 0 && (
+              <button
+                onClick={clearCompleted}
+                className="text-xs text-muted-foreground hover:text-foreground transition-colors px-3 py-1.5 rounded-lg hover:bg-muted"
+              >
+                清除已完成
+              </button>
+            )}
+            <button
+              onClick={onClose}
+              className="p-1.5 rounded-lg hover:bg-muted transition-colors"
+            >
+              <X className="h-4 w-4 text-muted-foreground" />
+            </button>
+          </div>
+        </div>
+
+        <div className="flex flex-1 min-h-0">
+          <div className="w-72 shrink-0 border-r border-border/20 flex flex-col min-h-0">
+            <div className="flex-1 min-h-0 overflow-y-auto">
+              {runningTasks.length > 0 && (
+                <div className="px-3 pt-3 pb-1">
+                  <p className="text-[10px] font-medium text-muted-foreground px-1 pb-1.5 uppercase tracking-wider">
+                    运行中 ({runningTasks.length})
+                  </p>
+                  {runningTasks.map((task) => (
+                    <TaskListItem
+                      key={task.id}
+                      label={task.label}
+                      subtitle={<ElapsedText task={task} />}
+                      icon={<Loader2 className="h-3.5 w-3.5 animate-spin text-blue-400 shrink-0" />}
+                      selected={selected?.type === "pipeline" && selected.taskId === task.id}
+                      onClick={() => setSelected({ type: "pipeline", taskId: task.id })}
+                    />
+                  ))}
+                </div>
+              )}
+
+              {completedTasks.length > 0 && (
+                <div className="px-3 pt-2 pb-1">
+                  <p className="text-[10px] font-medium text-muted-foreground px-1 pb-1.5 uppercase tracking-wider">
+                    当前会话 ({completedTasks.length})
+                  </p>
+                  {completedTasks.map((task) => (
+                    <TaskListItem
+                      key={task.id}
+                      label={task.label}
+                      subtitle={`${
+                        task.status === "done"
+                          ? "已完成"
+                          : task.status === "error"
+                            ? "出错"
+                            : "已取消"
+                      } · ${getElapsedStr(task)}`}
+                      icon={
+                        task.status === "done" ? (
+                          <CheckCircle2 className="h-3.5 w-3.5 text-green-400 shrink-0" />
+                        ) : task.status === "error" ? (
+                          <XCircle className="h-3.5 w-3.5 text-destructive shrink-0" />
+                        ) : (
+                          <Ban className="h-3.5 w-3.5 text-muted-foreground shrink-0" />
+                        )
+                      }
+                      selected={selected?.type === "pipeline" && selected.taskId === task.id}
+                      onClick={() => setSelected({ type: "pipeline", taskId: task.id })}
+                    />
+                  ))}
+                </div>
+              )}
+
+              <div className="px-3 pt-2 pb-3">
+                <p className="text-[10px] font-medium text-muted-foreground px-1 pb-1.5 uppercase tracking-wider flex items-center gap-1">
+                  <Clock className="h-3 w-3" />
+                  历史记录
+                </p>
+                {conversations.length === 0 && !historyLoading && (
+                  <p className="text-xs text-muted-foreground/60 px-1 py-3 text-center">
+                    暂无历史记录
+                  </p>
+                )}
+                {conversations.map((conversation) => {
+                  const isError =
+                    conversation.status === "error" ||
+                    conversation.status === "failed";
+                  const isDone =
+                    conversation.status === "completed" ||
+                    conversation.status === "done";
+                  const isCancelled = conversation.status === "cancelled";
+
+                  return (
+                    <TaskListItem
+                      key={conversation.id}
+                      label={conversation.title}
+                      subtitle={
+                        <>
+                          {conversation.agentType
+                            ? getAgentTypeName(conversation.agentType)
+                            : "对话"}
+                          {isError
+                            ? " · 出错"
+                            : isCancelled
+                              ? " · 已取消"
+                              : isDone
+                                ? " · 已完成"
+                                : ""}
+                          {conversation.createTime && conversation.lastMessageTime && (() => {
+                            const duration =
+                              new Date(conversation.lastMessageTime).getTime() -
+                              new Date(conversation.createTime).getTime();
+                            return duration > 0 ? ` · ${formatElapsed(duration)}` : "";
+                          })()}
+                        </>
+                      }
+                      icon={
+                        isError ? (
+                          <XCircle className="h-3.5 w-3.5 text-destructive shrink-0" />
+                        ) : isDone ? (
+                          <CheckCircle2 className="h-3.5 w-3.5 text-green-400 shrink-0" />
+                        ) : isCancelled ? (
+                          <Ban className="h-3.5 w-3.5 text-muted-foreground shrink-0" />
+                        ) : (
+                          <MessageSquare className="h-3.5 w-3.5 text-muted-foreground/50 shrink-0" />
+                        )
+                      }
+                      selected={
+                        selected?.type === "history" &&
+                        selected.conversation.id === conversation.id
+                      }
+                      onClick={() =>
+                        setSelected({ type: "history", conversation })
+                      }
+                    />
+                  );
+                })}
+
+                <div ref={listEndRef} className="h-1" />
+                {historyLoading && (
+                  <div className="flex items-center justify-center py-3">
+                    <Loader2 className="h-4 w-4 animate-spin text-muted-foreground" />
+                  </div>
+                )}
+              </div>
+            </div>
+          </div>
+
+          <div className="flex-1 min-w-0 min-h-0 bg-muted/10">
+            {selectedPipelineTask ? (
+              <PipelineDetailPanel task={selectedPipelineTask} />
+            ) : selectedConversation ? (
+              <HistoryDetailPanel conversation={selectedConversation} />
+            ) : (
+              <div className="flex items-center justify-center h-full text-muted-foreground">
+                <div className="text-center">
+                  <MessageSquare className="h-8 w-8 mx-auto mb-2 opacity-30" />
+                  <p className="text-sm">选择一个任务查看详情</p>
+                </div>
+              </div>
+            )}
+          </div>
+        </div>
+      </motion.div>
+    </>,
+    document.body
+  );
+}
